Refactored and new filter for `on` 🌂
